+++ std/src/console/cmds.rs
@@ -168,8 +168,8 @@ If a key press is available to be read, returns its name.  Otherwise, returns th
 The returned key matches its name, number, or symbol and maintains case.  In other words, \
 pressing the X key will return 'x' or 'X' depending on the SHIFT modifier.
 The following special keys are recognized: arrow keys (UP, DOWN, LEFT, RIGHT), backspace (BS), \
-end or CTRL+E (END), enter (ENTER), CTRL+D (EOF), escape (ESC), home or CTRL+A (HOME) and \
-CTRL+C (INT).
+end or CTRL+E (END), enter (ENTER), CTRL+D (EOF), escape (ESC), home or CTRL+A (HOME), \
+CTRL+C (INT) and tab (TAB).
 This function never blocks.  To wait for a key press, you need to explicitly poll the keyboard.  \
 For example, to wait until the escape key is pressed, you could do:
 k$ = \"\": WHILE k$ = <> \"ESC\": k = INKEY$(): SLEEP 0.01: WEND
@@ -209,6 +209,7 @@ impl Function for InKeyFunction {
             Some(Key::Home) => Value::Text("HOME".to_owned()),
             Some(Key::Interrupt) => Value::Text("INT".to_owned()),
             Some(Key::NewLine) => Value::Text("ENTER".to_owned()),
+            Some(Key::Tab) => Value::Text("TAB".to_owned()),
             Some(Key::Unknown(_)) => Value::Text("".to_owned()),
 
             None => Value::Text("".to_owned()),