+++ std/src/console/readline.rs
@@ -210,6 +210,10 @@ async fn read_line_interactive(
                 break;
             }
 
+            Key::Tab => {
+                // TODO(jmmv): Would be nice to have some form of auto-completion.
+            }
+
             // TODO(jmmv): Should do something smarter with unknown keys.
             Key::Unknown(_) => (),
         }
@@ -252,6 +256,7 @@ async fn read_line_raw(console: &mut dyn Console) -> io::Result<String> {
             Key::Eof => return Err(io::Error::new(io::ErrorKind::UnexpectedEof, "EOF")),
             Key::Interrupt => return Err(io::Error::new(io::ErrorKind::Interrupted, "Ctrl+C")),
             Key::NewLine => break,
+            Key::Tab => (),
             Key::Unknown(bad_input) => line += &bad_input,
         }
     }
@@ -842,6 +847,7 @@ mod tests {
             .add_output_bytes(b"not ")
             // -
             .add_key(Key::Escape)
+            .add_key(Key::Tab)
             // -
             .add_key_chars("affected")
             .add_output_bytes(b"affected")