+++ NEWS.md
@@ -50,6 +50,9 @@ for the time being.**
 *   Added support for the Home and End keys in the text editor and in text
     input operations.
 
+*   Improved indentation support in the editor by making the Backspace and
+    Tab keys aware of indentation.
+
 ## Changes in version 0.7.0
 
 **Released on 2021-07-03.**